Docker/requirements.txt
"numpy<2.0"
google-auth
google-cloud-bigquery-storage
google-cloud-bigquery
db-dtypes
pyarrow
google-cloud-storage
